ref : 2 digit after coma on circle and fix error input umur(age)

+++ internals/models/dataDiri.models.go
@@ -9,7 +9,7 @@ type DataDiri struct {
 	Nama        string
 	Foto        string
 	Email       string
-	Umur        uint8
+	Umur        int
 	NoTel       string
 	SPernikahan bool
 	Pendidikan  []DataSekolah
